Fleet commands get DI'd
