+++ FrEee.Gameplay/Commands/Designs/DesignCommandFactory.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FrEee.Gameplay.Commands.Designs;
 using FrEee.Gameplay.Commands.Notes;
+using FrEee.Gameplay.Commands.Orders;
 using FrEee.Objects.Civilization;
 using FrEee.Objects.Civilization.Orders;
 using FrEee.Objects.GameState;
 using FrEee.Objects.Vehicles;
 
-namespace FrEee.Gameplay.Commands.Orders;
+namespace FrEee.Gameplay.Commands.Designs;
 public class DesignCommandFactory
 	: IDesignCommandFactory
 {
