+++ FrEee.Gameplay/Commands/Notes/NoteCommandFactory.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FrEee.Gameplay.Commands.Notes;
+using FrEee.Gameplay.Commands.Orders;
 using FrEee.Objects.Civilization;
 using FrEee.Objects.Civilization.Orders;
 using FrEee.Objects.GameState;
 
-namespace FrEee.Gameplay.Commands.Orders;
+namespace FrEee.Gameplay.Commands.Notes;
 public class NoteCommandFactory
 	: INoteCommandFactory
 {
