+++ FrEee.Root/Configuration.cs
@@ -3,7 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using FrEee.Gameplay.Commands.Designs;
+using FrEee.Gameplay.Commands.Fleets;
+using FrEee.Gameplay.Commands.Notes;
 using FrEee.Gameplay.Commands.Orders;
+using FrEee.Gameplay.Commands.Orders.Fleets;
 using FrEee.Processes;
 using FrEee.Processes.Combat;
 using FrEee.Processes.Combat.Grid;
@@ -27,6 +31,7 @@ public static class Configuration
 
 		// gameplay
 		DI.RegisterSingleton<IDesignCommandFactory, DesignCommandFactory>();
+		DI.RegisterSingleton<IFleetCommandFactory, FleetCommandFactory>();
 		DI.RegisterSingleton<IOrderCommandFactory, OrderCommandFactory>();
 		DI.RegisterSingleton<INoteCommandFactory, NoteCommandFactory>();
